Refactor split tests into multiple files
